ai_call_krish/requirement.txt
# LangChain Core
langchain
langchain-core



# Environment Variable Management
python-dotenv

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai